feat: rename related to repository - localstorage_variable_api to variable_db - varialbe_api to variable_interface
